requirements.txt
matplotlib
opencv-python
tensorflow==2.13.1
numpy
tqdm
pdf2image
yapf
addict
pillow
tabulate
